CallbackManager

This module contains functions that handle dates and times in simulations. The functions in this module often call functions from Julia's Dates module.

CallbackManager API

ClimaUtilities.CallbackManager.HourlyCallback — Type

HourlyCallback{FT}

This is a callback type that triggers at intervals of 1h or multiple hours.

ClimaUtilities.CallbackManager.MonthlyCallback — Type

MonthlyCallback{FT}

This is a callback type that triggers at intervals of 1 month or multiple months.

Missing docstring.

Missing docstring for ClimaUtilities.CallbackManager.Monthly. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ClimaUtilities.CallbackManager.EveryTimestep. Check Documenter's build log for details.

ClimaUtilities.CallbackManager.to_datetime — Function

to_datetime(date)

Convert a DateTime-like object (e.g. DateTimeNoLeap) to a DateTime. We need this since some data files we use contain DateTimeNoLeap objects for dates, which can't be used for math with DateTimes. The DateTimeNoLeap type uses the Gregorian calendar without leap years, while the DateTime type uses Gregorian calendar with leap years.

For consistency, all input data files should have dates converted to DateTime before being used in a simulation.

This function is similar to reinterpret in CFTime.jl.

Arguments

date: DateTime-like object to be converted to DateTime

ClimaUtilities.CallbackManager.strdate_to_datetime — Function

strdate_to_datetime(strdate::String)

Convert from String ("YYYYMMDD") to Date format, required by the official AMIP input files.

ClimaUtilities.CallbackManager.datetime_to_strdate — Function

datetime_to_strdate(datetime::Dates.DateTime)

Convert from DateTime to String ("YYYYMMDD") format.

ClimaUtilities.CallbackManager.trigger_callback! — Function

trigger_callback!(callback, date_current)

If the callback is active and the current date is equal to or later than the "next call" reference date/time, call the callback function and increment the next call date based on the callback frequency. Otherwise, do nothing and leave the next call date unchanged.

Note that the collection of data in callback.data must match the types, number, and orderof arguments expected by callback.func.